+++ Revo.Core/Collections/MultiValueDictionary.cs
@@ -2,8 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.Serialization;
-using System.Text;
 
 namespace Revo.Core.Collections
 {