+++ Revo.Infrastructure/Repositories/EventSourcedAggregateStore.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using Revo.DataAccess.Entities;
 using Revo.Domain.Entities;
 using Revo.Domain.Entities.EventSourcing;
 using Revo.Infrastructure.EventSourcing;
@@ -20,27 +22,50 @@ namespace Revo.Infrastructure.Repositories
 
         public void Add<T>(T aggregate) where T : class, IAggregateRoot
         {
-            ((dynamic)eventSourcedRepository).Add((dynamic)aggregate);
+            CheckGenericType<T>();
+            eventSourcedRepository.Add((IEventSourcedAggregateRoot) aggregate);
         }
 
         public T Find<T>(Guid id) where T : class, IAggregateRoot
         {
-            return ((dynamic)eventSourcedRepository).Find<T>(id);
+            CheckGenericType<T>();
+            var aggregate = eventSourcedRepository.Find(id);
+            return CheckAggregate<T>(id, aggregate, false);
         }
 
-        public Task<T> FindAsync<T>(Guid id) where T : class, IAggregateRoot
+        public async Task<T> FindAsync<T>(Guid id) where T : class, IAggregateRoot
         {
-            return ((dynamic)eventSourcedRepository).FindAsync<T>(id);
+            CheckGenericType<T>();
+            var aggregate = await eventSourcedRepository.FindAsync(id);
+            return CheckAggregate<T>(id, aggregate, false);
         }
-
+        
+        public async Task<T[]> FindManyAsync<T>(params Guid[] ids) where T : class, IAggregateRoot
+        {
+            CheckGenericType<T>();
+            var aggregates = await eventSourcedRepository.FindManyAsync(ids);
+            return aggregates.Select(x => CheckAggregate<T>(x.Id, x, false)).ToArray();
+        }
+        
         public T Get<T>(Guid id) where T : class, IAggregateRoot
         {
-            return ((dynamic)eventSourcedRepository).Get<T>(id);
+            CheckGenericType<T>();
+            var aggregate = eventSourcedRepository.Get(id);
+            return CheckAggregate<T>(id, aggregate, false);
         }
 
-        public Task<T> GetAsync<T>(Guid id) where T : class, IAggregateRoot
+        public async Task<T> GetAsync<T>(Guid id) where T : class, IAggregateRoot
         {
-            return ((dynamic)eventSourcedRepository).GetAsync<T>(id);
+            CheckGenericType<T>();
+            var aggregate = await eventSourcedRepository.GetAsync(id);
+            return CheckAggregate<T>(id, aggregate, false);
+        }
+
+        public async Task<T[]> GetManyAsync<T>(params Guid[] ids) where T : class, IAggregateRoot
+        {
+            CheckGenericType<T>();
+            var aggregates = await eventSourcedRepository.GetManyAsync(ids);
+            return aggregates.Select(x => CheckAggregate<T>(x.Id, x, true)).ToArray();
         }
 
         public IEnumerable<IAggregateRoot> GetTrackedAggregates()
@@ -60,7 +85,37 @@ namespace Revo.Infrastructure.Repositories
 
         public void Remove<T>(T aggregate) where T : class, IAggregateRoot
         {
-            eventSourcedRepository.Remove((dynamic)aggregate);
+            CheckGenericType<T>();
+            eventSourcedRepository.Remove((IEventSourcedAggregateRoot)aggregate);
+        }
+
+        private void CheckGenericType<T>()
+        {
+            if (!typeof(IEventSourcedAggregateRoot).IsAssignableFrom(typeof(T)))
+            {
+                throw new InvalidOperationException($"Cannot use type {typeof(T).FullName} as an aggregate type with {this.GetType().Name} because it does not implement {nameof(IEventSourcedAggregateRoot)}");
+            }
+        }
+
+        private T CheckAggregate<T>(Guid id, IAggregateRoot aggregate, bool throwOnError) where T : class, IAggregateRoot
+        {
+            if (aggregate == null)
+            {
+                return null;
+            }
+
+            T typedAggregate = aggregate as T;
+            if (typedAggregate == null)
+            {
+                if (throwOnError)
+                {
+                    throw new EntityNotFoundException($"Aggregate root with ID '{id}' is not of requested type '{typeof(T).FullName}'");
+                }
+
+                return null;
+            }
+
+            return typedAggregate;
         }
     }
 }