+++ Tests/Revo.Infrastructure.Tests/Repositories/EventSourcedAggregateStoreTests.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Revo.Infrastructure.EventSourcing;
 using Revo.Infrastructure.Repositories;